feat(deps): upgraded github.com/alexfalkowski/go-service to v1.428.0
https://github.com/alexfalkowski/go-service/pull/1297
