+++ go.mod
@@ -6,7 +6,7 @@ toolchain go1.23.6
 
 require (
 	github.com/alexfalkowski/go-health v1.21.0
-	github.com/alexfalkowski/go-service v1.427.0
+	github.com/alexfalkowski/go-service v1.428.0
 	github.com/golang-migrate/migrate/v4 v4.18.2
 	go.opentelemetry.io/otel v1.34.0
 	go.opentelemetry.io/otel/trace v1.34.0
