+++ internal/health/module.go
@@ -2,8 +2,8 @@ package health
 
 import (
 	"github.com/alexfalkowski/go-service/health"
-	"github.com/alexfalkowski/migrieren/health/transport/grpc"
-	"github.com/alexfalkowski/migrieren/health/transport/http"
+	"github.com/alexfalkowski/migrieren/internal/health/transport/grpc"
+	"github.com/alexfalkowski/migrieren/internal/health/transport/http"
 	"go.uber.org/fx"
 )
