+++ internal/health/transport/http/observer.go
@@ -3,7 +3,7 @@ package http
 import (
 	"github.com/alexfalkowski/go-health/server"
 	"github.com/alexfalkowski/go-service/health/transport/http"
-	"github.com/alexfalkowski/migrieren/migrate"
+	"github.com/alexfalkowski/migrieren/internal/migrate"
 )
 
 // NewHealthObserver for HTTP.
