+++ internal/migrate/migrate.go
@@ -5,9 +5,9 @@ import (
 	"errors"
 
 	"github.com/alexfalkowski/go-service/meta"
-	"github.com/alexfalkowski/migrieren/migrate/migrator"
-	"github.com/alexfalkowski/migrieren/migrate/telemetry/logger"
-	"github.com/alexfalkowski/migrieren/migrate/telemetry/tracer"
+	"github.com/alexfalkowski/migrieren/internal/migrate/migrator"
+	"github.com/alexfalkowski/migrieren/internal/migrate/telemetry/logger"
+	"github.com/alexfalkowski/migrieren/internal/migrate/telemetry/tracer"
 	"github.com/golang-migrate/migrate/v4"
 	_ "github.com/golang-migrate/migrate/v4/database/pgx/v5" // need this for migrations to work.
 	_ "github.com/golang-migrate/migrate/v4/source/file"
